requirements.txt
Django==1.6.1
Markdown==2.3.1
Pillow==2.3.0
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-allauth==0.15.0
django-annoying==0.7.9
django-pure-pagination==0.2.1
django-toolbelt==0.0.1
gunicorn==18.0
oauthlib==0.6.1
postmarkup==1.2.0
psycopg2==2.5.2
pybbm==0.15.1
pystache==0.5.3
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
sorl-thumbnail==11.12
static==1.0.2
wsgiref==0.1.2
